Fix the usage of DRI_DRIVER_DIR variable.
The build configuration for minigbm was lacking the dri config.
This causes DRI_DRIVER_DIR to be undefined.
In addition, when the dri config is properly included, the
definition will include the quotation marks, making the
STRINGIZE step unnecessary.
To find the correct directory in the snap, we now prefix
DRI_DRIVER_DIR with /snap/chromium/current/gnome-platform

Fixes LP:2004586
Fixes LP:2006646
KIVU-90