Remove the profile lock when the kiosk starts.
Chromium creates a profile lock file to track whether or not a profile
is in-use to avoid profile corruption. Chromium has code to safely
cleaning this file [1], but part of that handling can involve manual
intervention from the user in the form of a pop-up window asking if it
is safe to unlock the profile. Manual intervention is undesirable on a
device acting as a kiosk, and because the kiosk is managed as a daemon
by snapd the profile should not be in-use.

[1] https://chromium.googlesource.com/chromium/src/+/master/chrome/browser/process_singleton_posix.cc

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>